app/requirements.txt
python-gvm==23.5.1
fastapi==0.100.0
python-multipart==0.0.6
uvicorn[standard]==0.23.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
redis[hiredis]==4.6.0